Conflicts:
	README.md
	coupling_components/solver_wrappers/openfoam/v41.py
	coupling_components/solver_wrappers/openfoam/v41/coconut_interFoam/coconut_interFoam.C
	solver_modules.py
	tests/solver_wrappers/openfoam/openfoam.py

+++ coupling_components/coupled_solvers/gauss_seidel.py
@@ -20,10 +20,11 @@ class CoupledSolverGaussSeidel(Component):
 
         self.parameters = parameters
         self.settings = parameters['settings']
-        self.init_time = time.time()
+        self.start_init_time = time.time()  # start of initialization
 
         # read parameters
         self.case_name = self.settings.get('case_name', 'case')  # case name
+        self.settings['case_name'] = self.case_name  # make sure a case name is present
         self.timestep_start_global = self.settings['timestep_start']  # time step for global calculation (restart)
         self.timestep_start_current = self.settings['timestep_start']  # time step start for this calculation (restart)
         self.restart = self.timestep_start_current != 0  # true if restart
@@ -40,7 +41,7 @@ class CoupledSolverGaussSeidel(Component):
         self.index_other = None
         for index in range(2):
             parameters = self.parameters['solver_wrappers'][index]
-            # add timestep_start and delta_t to solver_wrapper settings
+            # add timestep_start, delta_t and save_restart to solver_wrapper settings
             tools.pass_on_parameters(self.settings, parameters['settings'], ['timestep_start', 'delta_t',
                                                                              'save_restart'])
             self.solver_wrappers.append(create_instance(parameters))
@@ -58,7 +59,8 @@ class CoupledSolverGaussSeidel(Component):
         self.y = None  # input interface of solver 1
         self.iteration = None  # iteration
         self.solver_level = 0  # 0 is main solver (time step is printed)
-        self.start_time = None
+        self.init_time = None
+        self.start_run_time = None
         self.run_time = None
         self.run_time_previous = 0
         self.iterations = []
@@ -75,7 +77,7 @@ class CoupledSolverGaussSeidel(Component):
             self.residual = []
             self.info = None
 
-        self.debug = False  # save results each iteration including residual interfaces
+        self.debug = self.settings.get('debug', False)  # save results each iteration including residual interfaces
         if self.debug:
             self.complete_solution_r = None
 
@@ -96,11 +98,12 @@ class CoupledSolverGaussSeidel(Component):
         self.y = self.solver_wrappers[0].get_interface_output().copy()
         self.convergence_criterion.initialize()
         self.predictor.initialize(self.x)
-        self.start_time = time.time()
-        self.init_time = self.start_time - self.init_time
+        self.start_run_time = time.time()  # start of calculation
+        self.init_time = self.start_run_time - self.start_init_time  # duration of initialization
 
-        title = '╔' + 78 * '═' + f'╗\n║{self.case_name.upper():^78}║\n╚' + 78 * '═' + '╝\n'
-        tools.print_info(title)
+        if self.solver_level == 0:
+            title = '╔' + 78 * '═' + f'╗\n║{self.case_name.upper():^78}║\n╚' + 78 * '═' + '╝\n'
+            tools.print_info(title)
 
         # restart
         if self.restart:
@@ -193,7 +196,7 @@ class CoupledSolverGaussSeidel(Component):
             self.add_restart_data(output)
             with open(self.case_name + f'_restart_ts{self.time_step}.pickle', 'wb') as file:
                 pickle.dump(output, file)
-            if self.save_restart < 0:
+            if self.save_restart < 0 and self.time_step + self.save_restart > self.timestep_start_current:
                 try:
                     os.remove(self.case_name + f'_restart_ts{self.time_step + self.save_restart}.pickle')
                 except OSError:
@@ -211,7 +214,7 @@ class CoupledSolverGaussSeidel(Component):
     def output_solution_step(self):
         super().output_solution_step()
 
-        self.run_time = time.time() - self.start_time
+        self.run_time = time.time() - self.start_run_time  # duration of calculation
         if self.save_results != 0 and (self.time_step % self.save_results == 0 or
                                        (self.save_restart != 0 and self.time_step % self.save_restart == 0)):
             output = {'solution_x': self.complete_solution_x, 'solution_y': self.complete_solution_y,
@@ -235,10 +238,10 @@ class CoupledSolverGaussSeidel(Component):
             out = '╔' + 79 * '═' + '\n║\tSummary\n╠' + 79 * '═'
             tools.print_info(out)
 
-        self.print_summary()
-
         for component in self.components:
             component.finalize()
+
+        self.print_summary()
 
     def load_restart_data(self):
         restart_file_name = self.restart_case + f'_restart_ts{self.timestep_start_current}.pickle'
@@ -251,13 +254,13 @@ class CoupledSolverGaussSeidel(Component):
             for key in ('predictor', 'type'):
                 if self.parameters[key] != restart_data['parameters'][key]:
                     raise ValueError(f'Restart not possible because {key} changed in coupled solver')
-            self.add_restart_check(restart_data)
+            self.check_restart_data(restart_data)
             if self.delta_t != restart_data['delta_t']:
                 raise ValueError(f"Time step size has changed upon restart:\n\told: {restart_data['delta_t']}s"
                                  f"\n\tnew: {self.delta_t}s")
         return restart_data
 
-    def add_restart_check(self, restart_data):
+    def check_restart_data(self, restart_data):
         pass
 
     def add_restart_data(self, restart_data):
